move chart rendering functionality to new ChartRenderer class
