+++ src/main/java/org/mitre/synthea/engine/PhysiologySimulator.java
@@ -1,6 +1,5 @@
 package org.mitre.synthea.engine;
 
-import java.awt.Color;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -13,9 +12,7 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -25,16 +22,9 @@ import javax.xml.stream.XMLStreamException;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.math.ode.DerivativeException;
-import org.jfree.chart.ChartFactory;
-import org.jfree.chart.ChartUtils;
-import org.jfree.chart.JFreeChart;
-import org.jfree.chart.block.BlockBorder;
-import org.jfree.chart.plot.Plot;
-import org.jfree.chart.plot.PlotOrientation;
-import org.jfree.chart.plot.XYPlot;
-import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer;
-import org.jfree.data.xy.XYSeries;
-import org.jfree.data.xy.XYSeriesCollection;
+import org.mitre.synthea.helpers.ChartRenderer;
+import org.mitre.synthea.helpers.ChartRenderer.ChartConfig;
+import org.mitre.synthea.helpers.ChartRenderer.SeriesConfig;
 import org.sbml.jsbml.Model;
 import org.sbml.jsbml.SBMLDocument;
 import org.sbml.jsbml.SBMLException;
@@ -49,7 +39,6 @@ import org.simulator.math.odes.EulerMethod;
 import org.simulator.math.odes.GraggBulirschStoerSolver;
 import org.simulator.math.odes.HighamHall54Solver;
 import org.simulator.math.odes.MultiTable;
-import org.simulator.math.odes.MultiTable.Block.Column;
 import org.simulator.math.odes.RosenbrockSolver;
 import org.simulator.math.odes.RungeKutta_EventSolver;
 import org.simulator.sbml.SBMLinterpreter;
@@ -75,12 +64,6 @@ public class PhysiologySimulator {
   private final String[] modelFields;
   private final double[] modelDefaults;
   private final double simDuration;
-
-  /** Enumeration of supported chart types. **/
-  public enum ChartType {
-    SCATTER,
-    LINE
-  }
 
   /** POJO configuration for the simulation. **/
   public static class SimConfig {
@@ -148,171 +131,6 @@ public class PhysiologySimulator {
       this.charts = charts;
     }
     
-  }
-
-  /**
-   * POJO configuration for a chart.
-   **/
-  public static class ChartConfig {
-    /** Name of the image file to export. **/
-    private String filename;
-    /** User input for the type of chart to render. **/
-    private String type;
-    /** Chart title. **/
-    private String title;
-    /** Parameter to render on the x axis. **/
-    private String axisParamX;
-    /** X axis label. **/
-    private String axisLabelX;
-    /** Y axis label. **/
-    private String axisLabelY;
-    /** List of series configurations for this chart. **/
-    private List<SeriesConfig> series;
-    /** Simulation time in seconds to start charting points. **/
-    private double startTime;
-    /** Simulation time in seconds to end charting points. **/
-    private double endTime;
-    /** Chart width in pixels **/
-    private int width = 600;
-    /** Chart height in pixels **/
-    private int height = 300;
-    /** Whether the X Axis is visible **/
-    private boolean axisHiddenX = false;
-    /** CWhether the Y Axis is visible **/
-    private boolean axisHiddenY = false;
-    
-    public String getFilename() {
-      return filename;
-    }
-    
-    public void setFilename(String filename) {
-      this.filename = filename;
-    }
-    
-    public String getType() {
-      return type;
-    }
-    
-    public void setType(String type) {
-      this.type = type;
-    }
-    
-    public String getTitle() {
-      return title;
-    }
-    
-    public void setTitle(String title) {
-      this.title = title;
-    }
-    
-    public String getAxisParamX() {
-      return axisParamX;
-    }
-    
-    public void setAxisParamX(String axisParamX) {
-      this.axisParamX = axisParamX;
-    }
-    
-    public String getAxisLabelX() {
-      return axisLabelX;
-    }
-    
-    public void setAxisLabelX(String axisLabelX) {
-      this.axisLabelX = axisLabelX;
-    }
-    
-    public String getAxisLabelY() {
-      return axisLabelY;
-    }
-    
-    public void setAxisLabelY(String axisLabelY) {
-      this.axisLabelY = axisLabelY;
-    }
-    
-    public List<SeriesConfig> getSeries() {
-      return series;
-    }
-    
-    public void setSeries(List<SeriesConfig> series) {
-      this.series = series;
-    }
-    
-    public double getStartTime() {
-      return startTime;
-    }
-    
-    public void setStartTime(double startTime) {
-      this.startTime = startTime;
-    }
-    
-    public double getEndTime() {
-      return endTime;
-    }
-    
-    public void setEndTime(double endTime) {
-      this.endTime = endTime;
-    }
-
-    public int getWidth() {
-      return width;
-    }
-
-    public void setWidth(int width) {
-      this.width = width;
-    }
-
-    public int getHeight() {
-      return height;
-    }
-
-    public void setHeight(int height) {
-      this.height = height;
-    }
-
-    public boolean isAxisHiddenX() {
-      return axisHiddenX;
-    }
-
-    public void setAxisHiddenX(boolean axisHiddenX) {
-      this.axisHiddenX = axisHiddenX;
-    }
-
-    public boolean isAxisHiddenY() {
-      return axisHiddenY;
-    }
-
-    public void setAxisHiddenY(boolean axisHiddenY) {
-      this.axisHiddenY = axisHiddenY;
-    }
-    
-    
-  }
-
-  /**
-   * POJO configuration for a chart series.
-   */
-  public static class SeriesConfig {
-    /** Which parameter to plot on this series. **/
-    private String param;
-    /** Series label in the legend. **/
-    private String label;
-
-    public String getParam() {
-      return param;
-    }
-
-    public void setParam(String param) {
-      this.param = param;
-    }
-
-    public String getLabel() {
-      return label;
-    }
-
-    public void setLabel(String label) {
-      this.label = label;
-    }
-
   }
 
   static {
@@ -546,181 +364,6 @@ public class PhysiologySimulator {
   }
   
   /**
-   * Create a JFreeChart object based on values from a MultiTable.
-   * @param table MultiTable to retrieve values from
-   * @param config chart configuration options
-   */
-  private static JFreeChart createChart(MultiTable table, ChartConfig config) {
-    
-    // If there's only one series, and there's a title, hide the legend
-    
-    double lastTimePoint = table.getTimePoint(table.getRowCount() - 1);
-    
-    // Set the chart end time if not specified
-    if (config.getEndTime() == 0) {
-      config.setEndTime(lastTimePoint);
-    }
-    
-    // Check that the start time is valid
-    if (config.getStartTime() < 0) {
-      throw new IllegalArgumentException("Chart start time must not be negative");
-    }
-    
-    // Check the chart end time is valid
-    if (config.getEndTime() > lastTimePoint) {
-      throw new IllegalArgumentException("Invalid chart end time: " + config.getEndTime()
-          + " is greater than final time point " + lastTimePoint);
-    }
-    
-    // Check the time range is valid
-    if (config.getStartTime() > config.getEndTime()) {
-      throw new IllegalArgumentException("Invalid chart range: " + config.getStartTime()
-          + " to " + config.getEndTime());
-    }
-    
-    // Get the list of x values. Time is treated specially since it doesn't have a param identifier
-    boolean axisXIsTime = "time".equalsIgnoreCase(config.getAxisParamX());
-    List<Double> valuesX = new ArrayList<Double>(table.getRowCount());
-    double[] timePoints = table.getTimePoints();
-    Column colX = table.getColumn(config.getAxisParamX());
-    
-    // Check that the x axis identifier is valid
-    if (!axisXIsTime && colX == null) {
-      throw new IllegalArgumentException("Invalid X axis identifier: " + config.getAxisParamX());
-    }
-    
-    int startIndex = Arrays.binarySearch(timePoints, config.getStartTime());
-    int endIndex = Arrays.binarySearch(timePoints, config.getEndTime());
-    
-    // Add the table values to the list of x axis values within the provided time range
-    for (int i = startIndex; i < endIndex; i++) {
-      if (axisXIsTime) {
-        valuesX.add(timePoints[i]);
-      } else {
-        valuesX.add(colX.getValue(i));
-      }
-    }
-    
-    XYSeriesCollection dataset = new XYSeriesCollection();
-
-    // Add each series to the dataset
-    for (SeriesConfig seriesConfig : config.getSeries()) {
-      // If a label is not provided, use the parameter as the label
-      String seriesLabel = seriesConfig.getLabel();
-      if (seriesLabel == null) {
-        seriesLabel = seriesConfig.getParam();
-      }
-      
-      // don't auto-sort the series
-      XYSeries series = new XYSeries(seriesLabel, false);
-
-      Column col = table.getColumn(seriesConfig.getParam());
-      
-      // Check that the series identifier is valid
-      if (col == null) {
-        throw new IllegalArgumentException("Invalid series identifier: " + seriesConfig.getParam());
-      }
-
-      int indexX = 0;
-      for (int i = startIndex; i < endIndex; i++) {
-        series.add((double) valuesX.get(indexX++), col.getValue(i));
-      }
-
-      dataset.addSeries(series);
-    }
-    
-    // Instantiate our renderer to draw the chart
-    XYLineAndShapeRenderer renderer = new XYLineAndShapeRenderer();
-    JFreeChart chart;
-    
-    // Determine the appropriate Chart from the configuration options
-    switch (ChartType.valueOf(config.getType().toUpperCase())) {
-      default:
-      case LINE:
-        chart = ChartFactory.createXYLineChart(
-            config.getTitle(), 
-            config.getAxisLabelX(), 
-            config.getAxisLabelY(), 
-            dataset, 
-            PlotOrientation.VERTICAL,
-            true, 
-            true, 
-            false 
-        );
-        for (int i = 0; i < dataset.getSeriesCount(); i++) {
-          renderer.setSeriesShapesVisible(i, false);
-        }
-        break;
-      case SCATTER:
-        chart = ChartFactory.createScatterPlot(
-            config.getTitle(), 
-            config.getAxisLabelX(),
-            config.getAxisLabelY(),
-            dataset
-        );
-        break;
-    }
-    
-    // If there's only one series, and there's a chart title, the legend is unnecessary
-    if (config.getTitle() != null && !config.getTitle().isEmpty()
-        && config.getSeries().size() == 1) {
-      chart.removeLegend();
-    } else {
-      chart.getLegend().setFrame(BlockBorder.NONE);
-    }
-    
-    // Instantiate the plot and set some reasonable styles
-    // TODO eventually we can make these more configurable if desired
-    XYPlot plot = chart.getXYPlot();
-    
-    if (config.isAxisHiddenX()) {
-      plot.getDomainAxis().setVisible(false);
-    }
-    
-    if (config.isAxisHiddenY()) {
-      plot.getRangeAxis().setVisible(false);
-    }
-    
-    plot.setRenderer(renderer);
-    plot.setBackgroundPaint(Color.white);
-    plot.setRangeGridlinesVisible(true);
-    plot.setDomainGridlinesVisible(true);
-    
-    return chart;
-  }
-  
-  /**
-   * Draw a JFreeChart to an image based on values from a MultiTable.
-   * @param table MultiTable to retrieve values from
-   * @param config chart configuration options
-   */
-  public static void drawChart(MultiTable table, ChartConfig config) {
-    
-    JFreeChart chart = createChart(table, config);
-    
-    // Save the chart as a PNG image to the file system
-    try {
-      ChartUtils.saveChartAsPNG(new File(config.getFilename()), chart, config.getWidth(), config.getHeight());
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-  }
-  
-  /**
-   * Draw a JFreeChart to a base64 encoded image based on values from a MultiTable.
-   * @param table MultiTable to retrieve values from
-   * @param config chart configuration options
-   * @throws IOException 
-   */
-  public static String drawBase64Chart(MultiTable table, ChartConfig config) throws IOException {
-    
-    JFreeChart chart = createChart(table, config);
-
-    byte[] imgBytes = ChartUtils.encodeAsPNG(chart.createBufferedImage(config.getWidth(), config.getHeight()));
-    return new String(Base64.getEncoder().encode(imgBytes));
-  }
-  
-  /**
    * Retrieves the default value for a model parameter.
    * @param param parameter to search for
    * @return initial value
@@ -807,7 +450,7 @@ public class PhysiologySimulator {
           }
           chartConfig.setFilename(Paths.get(outputDir.toString(),
               chartConfig.getFilename()).toString());
-          PhysiologySimulator.drawChart(results, chartConfig);
+          ChartRenderer.drawChartAsFile(results, chartConfig);
         }
       }
